edlab-voicing/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
azure-cognitiveservices-speech==1.41.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
edge-tts==7.0.0
frozenlist==1.5.0
gTTS==2.5.4
idna==3.10
multidict==6.1.0
propcache==0.2.1
requests==2.32.3
srt==3.5.3
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
